vehicle-vision-service/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.6
numpy==1.26.0
pillow==10.1.0
opencv-python==4.8.1.78
tensorflow==2.14.0
pydantic==2.4.2
python-dotenv==1.0.0
